Rename tests to match file names.
